apps/data-pipeline/pyproject.toml
[tool.coverage.run]
branch = true
source = ["data_pipeline"]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../coverage/apps/data-pipeline/html' --cov-report xml:'../../coverage/apps/data-pipeline/coverage.xml' --html='../../reports/apps/data-pipeline/unittests/html/index.html' --junitxml='../../reports/apps/data-pipeline/unittests/junit.xml'"

[tool.poetry]
name = "data-pipeline"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = []
license = 'Proprietary'
readme = 'README.md'

[[tool.poetry.packages]]
include = "data_pipeline"

[tool.poetry.dependencies]
python = "3.12.7"
dagster = "^1.10.4"
polars = "^1.17.1"
dagster-polars = "^0.26.4"
dagster-cloud = "^1.10.4"
faiss-cpu = "^1.9.0.post1"
networkx = "^3.4.2"
json-repair = "^0.30.3"
universal-pathlib = "^0.2.5"
azure-storage-blob = "^12.24.0"
azure-storage-queue = "^12.12.0"
adlfs = "^2024.7.0"
scikit-learn = "^1.6.0"
pandas = "^2.2.3"
plotext = "^5.3.2"
umap-learn = "^0.5.7"
openai = "^1.60.0"
psycopg = { version = "^3.2.4", extras = ["binary"] }
google-auth = "^2.38.0"
hdbscan = "^0.8.40"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.3.2"
ruff = "^0.9.5"
pytest = "7.3.1"
pytest-sugar = "0.9.7"
pytest-cov = "4.1.0"
pytest-html = "3.2.0"
dagster-webserver = "^1.10.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.dagster]
module_name = "data_pipeline"


[tool.ruff]
exclude = [".ruff_cache", ".svn", ".tox", ".venv", "dist"]

line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I",
]
ignore = ["E501"]

fixable = ["ALL"]
unfixable = []
